Merge the Node 18 upgrade changes

See merge request gnome/oil_database/oil_database!18

+++ web_api/sync_with_production.sh
@@ -1,53 +1,42 @@
 #!/bin/bash
 
 # Script to backup oil database and set it to production branch
-echo "In sync_with_production.sh script"
 
-#set -e  # exit script if any commands fail
+echo "Inside Script: $0"
+set -e  # exit script if any commands fail
 set -x  # echo commands to stdout
 
-function branch_exists_remotely() {
-    local branch=${1}
-    local exists_in_remote=$(git ls-remote --heads origin ${branch})
-
-    if [[ -n ${exists_in_remote} ]]
-    then
-        echo 0
-    else
-        echo 1
-    fi
-}
-
+SCRIPT_DIR=$( cd -- "$( dirname -- "${0}" )" &> /dev/null && pwd )
+source ${SCRIPT_DIR}/git_functions.sh
 
 cd /data/git_repos/noaa-oil-data/
 
 git config user.email "adios-script@noaa.gov"
 git config user.name "Adios Automated Script"
 
-git status
 
-# backup the data in the database
-
-BRANCH=`date +'%Y/%m/%d_%H;%M'`
-
-# this will update the local list of remote branches
+# this will update the local list of remote branches, as they may change
+# when a merge request removes a branch and squashes commits.
 git remote update origin --prune
 
+git status
 
-# The local under_review branch might be in a different state as the remote
-# This might happen if a merge request removes a branch and squashes commits.
-git checkout production
-git pull
-
-git branch -D under_review
-
-if [[ $(branch_exists_remotely under_review) -eq "0" ]]
+if [[ $(branch_exists_remotely under_review) -eq "0" ]];
 then
     echo "The remote under_review branch exists."
-    git checkout under_review
+    BRANCH=under_review
 else
-    echo "The remote under_review branch does not exist.  Use production."
+    echo "The remote under_review branch does not exist.  Use a side branch."
+    BRANCH=`date +'%Y/%m/%d_%H;%M'`
 fi
+
+git checkout production
+git pull -s recursive -X theirs --no-edit
+
+# The local under_review branch might be in a different state as the remote
+# This might happen if a merge request on GitLab removes a branch and
+# squashes commits.  So we remove the local under_review branch
+git branch -D under_review || echo "local under_review not found. This is ok."
 
 git checkout -b $BRANCH
 
@@ -57,18 +46,20 @@ git add -A
 
 if ! git diff-index --quiet HEAD --
 then
-    # changes made
+    # changes made in the MongoDB will be saved in the branch that we
+    # decided on.
     git commit -a -m "Save from Running Editable Server"
     git push --set-upstream origin $BRANCH
 fi
 
-# update database with latest production
+# Update the MongoDB with the GitLab's current production branch
 git checkout production
 git pull
 adios_db_restore --config /config/config_oil_db.ini
 echo "Database is ready to be edited ..."
 
-# delete the under_review branch
+# Since we are syncing with the production branch, the the under_review branch
+# changes will be dropped
 git branch -D under_review
 git push origin --delete under_review
 
@@ -77,4 +68,8 @@ git checkout production
 git checkout -b under_review
 git push --set-upstream origin under_review
 
-git branch -D $BRANCH  # clean up our local branch
+if [[ "$BRANCH" != "under_review" ]];
+then
+    echo "cleaning up the side-branch $BRANCH"
+    git branch -D $BRANCH  # clean up our local branch
+fi
